Add files via upload
To run the program, download this ipynb file, and upload it into jupyter notebook. Open it there, and then just run the file. When you are uploading an excel file with a CGM profile, make sure you have a column titled 'Timestamp' and another titled 'Glucose'. Then use the dropdown boxes and the slider to annotate across the days. Press the key 'u' on your keyboard to undo your last click, and finally download the file with the saved annotations.